requirements.txt
python-dotenv~=1.0.0
aio-pika==9.4.2
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
exceptiongroup==1.0.0rc9
filelock==3.8.0
h11==0.14.0
idna==3.4
lxml==5.2.2
numpy==1.26.4
ollama==0.1.5
outcome==1.2.0
packaging==21.3
pdfminer==20191125
pdfminer.six==20220524
pika==1.3.1
psycopg2-binary==2.9.3
pycparser==2.21
pycryptodome==3.15.0
pyparsing==3.0.9
PySocks==1.7.1
PyYAML==6.0.1
regex==2022.9.13
requests==2.28.1
selenium==4.5.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.8.0
urllib3==1.26.12
wsproto==1.2.0
